update files from TA

+++ podem/miscell.h
@@ -33,12 +33,12 @@ struct FAULT {
     short activate;
     short detect;         /* detection flag */
     short test_tried;     /* flag to indicate test is being tried */
-    short eqv_fault_num;  /* number of equivalent faults */
+    int eqv_fault_num;    /* number of equivalent faults */
     fptr pnext;           /* pointer to the next element in the list */
     fptr pnext_undetect;  /* pointer to next undetected fault */
-    short to_swlist;      /* index to the sort_wlist[] */ 
+    int to_swlist;        /* index to the sort_wlist[] */ 
     int fault_no;         /* fault index */
-    int detect_num;       /* number of detection */
+    int detect_num;       /* detected number */
 };
 
 fptr first_fault;         /* pointer to the first element of the fault list */
